Moved views from users to enterprise app
