+++ corehq/apps/users/urls.py
@@ -25,8 +25,6 @@ from .views import (
     test_httpdigest,
     undo_remove_web_user,
     verify_phone_number,
-    delete_domain_permission_mirror,
-    create_domain_permission_mirror,
     download_web_users,
     DownloadWebUsersStatusView,
     WebUserUploadJobPollView,
@@ -114,11 +112,6 @@ urlpatterns = [
         WebUserUploadStatusView.as_view(), name=WebUserUploadStatusView.urlname),
     url(r'^web/upload/poll/(?P<download_id>(?:dl-)?[0-9a-fA-Z]{25,32})/$', WebUserUploadJobPollView.as_view(),
         name=WebUserUploadJobPollView.urlname),
-    # TODO: move these 2
-    url(r'^enterprise/delete_domain_permission_mirror/(?P<mirror>[ \w-]+)/$', delete_domain_permission_mirror,
-        name='delete_domain_permission_mirror'),
-    url(r'^enterprise/create_domain_permission_mirror/$', create_domain_permission_mirror,
-        name='create_domain_permission_mirror'),
     url(r'^join/(?P<uuid>[ \w-]+)/$', accept_invitation, name='domain_accept_invitation'),
     url(r'^roles/$', ListRolesView.as_view(), name=ListRolesView.urlname),
     url(r'^roles/save/$', post_user_role, name='post_user_role'),
